chore: use NDataTable for column list
